+++ src/main/java/services/task_presentation/TaskGetter.java
@@ -6,14 +6,12 @@ import data_gateway.TodoListManager;
 import java.util.List;
 import java.util.Map;
 
-public class TaskGetter implements TodoListDisplayBoundary {
+public class TaskGetter implements TodoListRequestBoundary {
 
     private final TodoListManager todoListManager;
-    private final TodoListPresenter taskPresenter;
 
-    public TaskGetter(TodoListManager todoListManager, TodoListPresenter taskPresenter) {
+    public TaskGetter(TodoListManager todoListManager) {
         this.todoListManager = todoListManager;
-        this.taskPresenter = taskPresenter;
     }
 
     /**
@@ -21,6 +19,7 @@ public class TaskGetter implements TodoListDisplayBoundary {
      * @param id id of a task
      * @return the corresponding task as a TaskInfo
      */
+    @Override
     public TaskInfo getTaskById(Long id) {
         Map<Long, List<TaskReader>> taskMap = todoListManager.getAllTasks();
         for (List<TaskReader> todoListTasks : taskMap.values())
@@ -32,24 +31,4 @@ public class TaskGetter implements TodoListDisplayBoundary {
         return null;
     }
 
-    /**
-     * Get tasks from the database through data gateway and
-     * pass the tasks as DTOs to the presenter to present all tasks.
-     */
-    public void presentAllTasks() {
-        Map<Long, List<TaskReader>> taskReaders = todoListManager.getAllTasks();
-        TodoListsInfo todoListInfo = new TodoListInfoFromTaskReaders(taskReaders);
-        taskPresenter.presentTasks(todoListInfo);
-    }
-
-    /**
-     * Get tasks from the database through data gateway and
-     * pass the tasks as DTOs to the presenter to present all tasks in an ordered list.
-     * @return a mapping of task's position in the presented list and id
-     */
-    public Map<Integer, Long> presentAllTasksForUserSelection() {
-        Map<Long, List<TaskReader>> taskReaders = todoListManager.getAllTasks();
-        TodoListsInfo todoListInfo = new TodoListInfoFromTaskReaders(taskReaders);
-        return taskPresenter.presentTasksForUserSelection(todoListInfo);
-    }
 }
