+++ src/main/java/services/task_presentation/TodoListDisplayBoundary.java
@@ -1,5 +1,9 @@
 package services.task_presentation;
 
+import java.util.Map;
+
 public interface TodoListDisplayBoundary {
+    void presentAllTasks();
+    Map<Integer, Long> presentAllTasksForUserSelection();
 
 }
