+++ src/main/java/services/task_presentation/TodoListRequestBoundary.java
@@ -1,0 +1,5 @@
+package services.task_presentation;
+
+public interface TodoListRequestBoundary {
+    TaskInfo getTaskById(Long id);
+}
